Organize EMR classes in Amazon provider (#20160)
Organize EMR classes in Amazon provider (#20160)
