+++ airflow/providers/amazon/aws/sensors/emr_containers.py
@@ -14,80 +14,32 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-import sys
-from typing import Any, Optional
-
-if sys.version_info >= (3, 8):
-    from functools import cached_property
-else:
-    from cached_property import cached_property
-
-from airflow.exceptions import AirflowException
-from airflow.providers.amazon.aws.hooks.emr_containers import EMRContainerHook
-from airflow.sensors.base import BaseSensorOperator
 
 
-class EMRContainerSensor(BaseSensorOperator):
+"""This module is deprecated. Please use :mod:`airflow.providers.amazon.aws.sensors.emr`."""
+
+import warnings
+
+from airflow.providers.amazon.aws.sensors.emr import EmrBaseSensor, EmrContainerSensor  # noqa
+
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.amazon.aws.sensors.emr`.",
+    DeprecationWarning,
+    stacklevel=2,
+)
+
+
+class EMRContainerSensor(EmrContainerSensor):
     """
-    Asks for the state of the job run until it reaches a failure state or success state.
-    If the job run fails, the task will fail.
-
-    :param job_id: job_id to check the state of
-    :type job_id: str
-    :param max_retries: Number of times to poll for query state before
-        returning the current state, defaults to None
-    :type max_retries: int
-    :param aws_conn_id: aws connection to use, defaults to 'aws_default'
-    :type aws_conn_id: str
-    :param poll_interval: Time in seconds to wait between two consecutive call to
-        check query status on athena, defaults to 10
-    :type poll_interval: int
+    This class is deprecated.
+    Please use  :class:`airflow.providers.amazon.aws.sensors.emr.EmrContainerSensor`.
     """
 
-    INTERMEDIATE_STATES = (
-        "PENDING",
-        "SUBMITTED",
-        "RUNNING",
-    )
-    FAILURE_STATES = (
-        "FAILED",
-        "CANCELLED",
-        "CANCEL_PENDING",
-    )
-    SUCCESS_STATES = ("COMPLETED",)
-
-    template_fields = ['virtual_cluster_id', 'job_id']
-    template_ext = ()
-    ui_color = '#66c3ff'
-
-    def __init__(
-        self,
-        *,
-        virtual_cluster_id: str,
-        job_id: str,
-        max_retries: Optional[int] = None,
-        aws_conn_id: str = 'aws_default',
-        poll_interval: int = 10,
-        **kwargs: Any,
-    ) -> None:
+    def __init__(self, **kwargs):
+        warnings.warn(
+            """This class is deprecated.
+           Please use `airflow.providers.amazon.aws.sensors.emr.EmrContainerSensor`.""",
+            DeprecationWarning,
+            stacklevel=2,
+        )
         super().__init__(**kwargs)
-        self.aws_conn_id = aws_conn_id
-        self.virtual_cluster_id = virtual_cluster_id
-        self.job_id = job_id
-        self.poll_interval = poll_interval
-        self.max_retries = max_retries
-
-    def poke(self, context: dict) -> bool:
-        state = self.hook.poll_query_status(self.job_id, self.max_retries, self.poll_interval)
-
-        if state in self.FAILURE_STATES:
-            raise AirflowException('EMR Containers sensor failed')
-
-        if state in self.INTERMEDIATE_STATES:
-            return False
-        return True
-
-    @cached_property
-    def hook(self) -> EMRContainerHook:
-        """Create and return an EMRContainerHook"""
-        return EMRContainerHook(self.aws_conn_id, virtual_cluster_id=self.virtual_cluster_id)
