+++ tests/providers/amazon/aws/operators/test_emr_modify_cluster.py
@@ -23,7 +23,7 @@ import pytest
 
 from airflow.exceptions import AirflowException
 from airflow.models.dag import DAG
-from airflow.providers.amazon.aws.operators.emr_modify_cluster import EmrModifyClusterOperator
+from airflow.providers.amazon.aws.operators.emr import EmrModifyClusterOperator
 from airflow.utils import timezone
 
 DEFAULT_DATE = timezone.datetime(2017, 1, 1)
