+++ tests/providers/amazon/aws/sensors/test_emr_job_flow.py
@@ -24,7 +24,7 @@ import pytest
 from dateutil.tz import tzlocal
 
 from airflow.exceptions import AirflowException
-from airflow.providers.amazon.aws.sensors.emr_job_flow import EmrJobFlowSensor
+from airflow.providers.amazon.aws.sensors.emr import EmrJobFlowSensor
 
 DESCRIBE_CLUSTER_STARTING_RETURN = {
     'Cluster': {
